xserver libX11 fontconfig pixman putty pthreads xkeyboard-config git update Sep 18 2021
xserver          commit db9e9d45e8ba73510f11eb9e534c176102f6623e
xkeyboard-config commit 8c99663748e9af73e2fd0e63c7bcc82c4b3dc7f9
libX11           commit e92efc63acd7b377faa9e534f4bf52aaa86be2a9
pixman           commit eadb82866b0f6a326a61c36f60e5c2be8f7479af
fontconfig       commit 6e414d61c77b2e6c08d0694ae61b848ee836b7eb
putty            commit d489c64f48d50a17e688edc9fee4973790f8aca7
pthreads         commit 3b9e0c6e46324bd15df111bcc3560fe0ebcbd708

+++ pthreads/tests/exit3.c
@@ -39,7 +39,7 @@
 void *
 func(void * arg)
 {
-	int failed = (int) arg;
+	size_t failed = (size_t) arg;
 
 	pthread_exit(arg);
 
@@ -47,7 +47,7 @@ func(void * arg)
         /*
          * assert(0) in a way to prevent warning or optimising away.
          */
-	assert(failed - (int) arg);
+	assert(failed - (size_t) arg);
 
 	return NULL;
 }
